+++ server/src/main/java/com/ones/admin/system/ApiResourceService.java
@@ -283,7 +283,7 @@ public class ApiResourceService {
             SystemMenuMapper menuMapper,
             SystemApiManifestSnapshotMapper manifestSnapshotMapper,
             ObjectMapper objectMapper,
-            @Value("${ones.version:v0.0.46}") String applicationVersion
+            @Value("${ones.version:v0.0.47}") String applicationVersion
     ) {
         this.requestMappingHandlerMapping = requestMappingHandlerMapping;
         this.permissionMapper = permissionMapper;